pyproject.toml
[project]
name = "loominary"
version = "0.1.0"
description = "Podcast transcription tool using Spotify, Whisper, and DuckDB"
requires-python = ">=3.10"
dependencies = [
    "spotipy>=2.24.0",
    "feedparser>=6.0.11",
    "requests>=2.31.0",
    "faster-whisper>=1.0.0",
    "duckdb>=0.10.0",
    "pandas>=2.2.0",
    "google-api-python-client>=2.120.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "rich>=13.7.0",
    "questionary>=2.0.1",
    "python-dotenv>=1.0.1",
    "python-dateutil>=2.9.0",
    "openai-whisper>=20231117",
]

[project.scripts]
loominary = "main:main"

[tool.uv]
dev-dependencies = []
